Separate batch transcription backends
