+++ crates/listener2-core/src/batch/simple.rs
@@ -1,1327 +1,21 @@
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::time::{Duration, Instant};
-
-use owhisper_client::{
-    AdapterKind, AnarlogAdapter, AquaVoiceAdapter, ArgmaxAdapter, AssemblyAIAdapter,
-    AwsTranscribeAdapter, AzureSpeechAdapter, BatchSttAdapter, CartesiaAdapter, CohereAdapter,
-    DeepgramAdapter, ElevenLabsAdapter, FireworksAdapter, GladiaAdapter, GoogleCloudAdapter,
-    GroqAdapter, MistralAdapter, OpenAIAdapter, PyannoteAdapter, RevAiAdapter, SonioxAdapter,
-    SpeechmaticsAdapter, TogetherAdapter, XaiAdapter,
-};
-use owhisper_interface::batch_stream::BatchStreamEvent;
-use tracing::Instrument;
-
-use anlg_audio_chunking::AudioChunk;
-use anlg_audio_utils::Source;
-use anlg_transcribe_core::{
-    TARGET_SAMPLE_RATE, channel_duration_sec, chunk_channel_audio, split_resampled_channels,
-};
-
-use super::{BatchParams, BatchRunMode, BatchRunOutput, format_user_friendly_error, session_span};
-use crate::{BatchEvent, BatchRuntime};
-
-const SONIQO_PARAKEET_MAX_CHUNK_SAMPLES: usize = TARGET_SAMPLE_RATE as usize * 59 / 2;
-const SONIQO_PROGRESS_PLANNED: f64 = 0.05;
-const SONIQO_PROGRESS_RANGE: f64 = 0.90;
-const SONIQO_PROGRESS_MAX: f64 = 0.95;
-const SONIQO_DIRECT_MIC_MIN_RMS: f64 = 0.0008;
-const DIRECT_BATCH_TIMEOUT_FLOOR: Duration = Duration::from_secs(15 * 60);
-const DIRECT_BATCH_TIMEOUT_CEILING: Duration = Duration::from_secs(6 * 60 * 60);
-const DIRECT_BATCH_TIMEOUT_BUFFER: Duration = Duration::from_secs(5 * 60);
-const DIRECT_BATCH_AUDIO_DURATION_MULTIPLIER: u32 = 2;
-const ANARLOG_PROXY_MAX_AUDIO_BYTES: u64 = 512 * 1024 * 1024;
-
-enum PreparedBatchUpload {
-    Original(PathBuf),
-    Compressed {
-        _temp_dir: tempfile::TempDir,
-        path: PathBuf,
-    },
-}
-
-impl PreparedBatchUpload {
-    fn path(&self) -> &Path {
-        match self {
-            Self::Original(path) | Self::Compressed { path, .. } => path,
-        }
-    }
-}
-
-macro_rules! dispatch_batch {
-    ($ak:expr, $params:expr, $lp:expr,
-     { $($var:ident => $adapter:ty),+ $(,)? },
-     unsupported: [$($unsup:ident),* $(,)?]
-    ) => {
-        match $ak {
-            $(AdapterKind::$var => {
-                run_direct_batch::<$adapter>(&AdapterKind::$var.to_string(), $params, $lp).await
-            })+
-            $(AdapterKind::$unsup => {
-                Err(crate::BatchFailure::DirectBatchUnsupported {
-                    provider: AdapterKind::$unsup.to_string(),
-                }.into())
-            })*
-        }
-    };
-}
-
-pub(super) async fn run_direct_batch_for_adapter_kind(
-    adapter_kind: AdapterKind,
-    params: BatchParams,
-    listen_params: owhisper_interface::ListenParams,
-) -> crate::Result<BatchRunOutput> {
-    if adapter_kind == AdapterKind::Anarlog {
-        return run_anarlog_batch(params, listen_params).await;
-    }
-
-    dispatch_batch!(adapter_kind, params, listen_params, {
-        Argmax => ArgmaxAdapter,
-        Cartesia => CartesiaAdapter,
-        Deepgram => DeepgramAdapter,
-        Soniox => SonioxAdapter,
-        AssemblyAI => AssemblyAIAdapter,
-        Fireworks => FireworksAdapter,
-        OpenAI => OpenAIAdapter,
-        Gladia => GladiaAdapter,
-        ElevenLabs => ElevenLabsAdapter,
-        Pyannote => PyannoteAdapter,
-        Mistral => MistralAdapter,
-        Anarlog => AnarlogAdapter,
-        AquaVoice => AquaVoiceAdapter,
-        Cohere => CohereAdapter,
-        AwsTranscribe => AwsTranscribeAdapter,
-        AzureSpeech => AzureSpeechAdapter,
-        GoogleCloud => GoogleCloudAdapter,
-        Groq => GroqAdapter,
-        RevAi => RevAiAdapter,
-        Speechmatics => SpeechmaticsAdapter,
-        Together => TogetherAdapter,
-        Xai => XaiAdapter,
-    }, unsupported: [DashScope])
-}
-
-async fn run_anarlog_batch(
-    mut params: BatchParams,
-    listen_params: owhisper_interface::ListenParams,
-) -> crate::Result<BatchRunOutput> {
-    let upload =
-        prepare_anarlog_batch_upload(&params.file_path, ANARLOG_PROXY_MAX_AUDIO_BYTES).await?;
-    params.file_path = upload.path().to_string_lossy().into_owned();
-    run_direct_batch::<AnarlogAdapter>(&AdapterKind::Anarlog.to_string(), params, listen_params)
-        .await
-}
-
-async fn prepare_anarlog_batch_upload(
-    file_path: &str,
-    max_bytes: u64,
-) -> crate::Result<PreparedBatchUpload> {
-    let source_path = PathBuf::from(file_path);
-    let source_size = tokio::fs::metadata(&source_path).await?.len();
-    if source_size <= max_bytes {
-        return Ok(PreparedBatchUpload::Original(source_path));
-    }
-
-    let is_wav = source_path
-        .extension()
-        .and_then(|extension| extension.to_str())
-        .is_some_and(|extension| extension.eq_ignore_ascii_case("wav"));
-    if !is_wav {
-        return Err(crate::BatchFailure::DirectRequestFailed {
-            provider: AdapterKind::Anarlog.to_string(),
-            message:
-                "This recording is too large for cloud transcription. Convert it to MP3 and try again."
-                    .to_string(),
-        }
-        .into());
-    }
-
-    let temp_dir = tempfile::tempdir().map_err(|error| {
-        tracing::error!(%error, "large_batch_audio_temp_dir_failed");
-        crate::BatchFailure::DirectRequestFailed {
-            provider: AdapterKind::Anarlog.to_string(),
-            message: "Anarlog couldn't prepare this large recording for transcription.".to_string(),
-        }
-    })?;
-    let encoded_path = temp_dir.path().join("audio.mp3");
-    let encode_source = source_path.clone();
-    let encode_target = encoded_path.clone();
-    tokio::task::spawn_blocking(move || anlg_mp3::encode_wav(&encode_source, &encode_target))
-        .await
-        .map_err(|error| {
-            tracing::error!(%error, "large_batch_audio_encode_task_failed");
-            crate::BatchFailure::DirectRequestFailed {
-                provider: AdapterKind::Anarlog.to_string(),
-                message: "Anarlog couldn't prepare this large recording for transcription."
-                    .to_string(),
-            }
-        })?
-        .map_err(|error| {
-            tracing::error!(%error, "large_batch_audio_encode_failed");
-            crate::BatchFailure::DirectRequestFailed {
-                provider: AdapterKind::Anarlog.to_string(),
-                message: "Anarlog couldn't prepare this large recording for transcription."
-                    .to_string(),
-            }
-        })?;
-
-    let encoded_size = tokio::fs::metadata(&encoded_path).await?.len();
-    if encoded_size > max_bytes {
-        return Err(crate::BatchFailure::DirectRequestFailed {
-            provider: AdapterKind::Anarlog.to_string(),
-            message:
-                "This recording is too large for cloud transcription. Split it into smaller files and try again."
-                    .to_string(),
-        }
-        .into());
-    }
-
-    tracing::info!(
-        source_size,
-        encoded_size,
-        "large_batch_audio_compressed_for_upload"
-    );
-
-    Ok(PreparedBatchUpload::Compressed {
-        _temp_dir: temp_dir,
-        path: encoded_path,
-    })
-}
-
-async fn run_direct_batch<A: BatchSttAdapter>(
-    provider: &str,
-    params: BatchParams,
-    listen_params: owhisper_interface::ListenParams,
-) -> crate::Result<BatchRunOutput> {
-    let timeout = direct_batch_timeout(&params.file_path);
-    run_direct_batch_with_timeout::<A>(provider, params, listen_params, timeout).await
-}
-
-async fn run_direct_batch_with_timeout<A: BatchSttAdapter>(
-    provider: &str,
-    params: BatchParams,
-    listen_params: owhisper_interface::ListenParams,
-    timeout: Duration,
-) -> crate::Result<BatchRunOutput> {
-    let span = session_span(&params.session_id);
-
-    async {
-        let client = owhisper_client::BatchClient::<A>::builder()
-            .api_base(params.base_url.clone())
-            .api_key(params.api_key.clone())
-            .params(listen_params)
-            .build();
-
-        tracing::debug!("transcribing file: {}", params.file_path);
-        let response =
-            match tokio::time::timeout(timeout, client.transcribe_file(&params.file_path)).await {
-                Ok(Ok(response)) => response,
-                Ok(Err(err)) => {
-                    let raw_error = format!("{err:?}");
-                    let message = format_user_friendly_error(&raw_error);
-                    tracing::error!(
-                        error = %raw_error,
-                        anarlog.error.user_message = %message,
-                        "batch transcription failed"
-                    );
-                    return Err(crate::BatchFailure::DirectRequestFailed {
-                        provider: provider.to_string(),
-                        message,
-                    }
-                    .into());
-                }
-                Err(_) => {
-                    tracing::error!(
-                        timeout_seconds = timeout.as_secs(),
-                        "batch transcription timed out"
-                    );
-                    return Err(crate::BatchFailure::DirectRequestTimedOut {
-                        provider: provider.to_string(),
-                        timeout_seconds: timeout.as_secs(),
-                    }
-                    .into());
-                }
-            };
-        tracing::info!("batch transcription completed");
-
-        Ok(BatchRunOutput {
-            session_id: params.session_id,
-            mode: BatchRunMode::Direct,
-            response,
-        })
-    }
-    .instrument(span)
-    .await
-}
-
-fn direct_batch_timeout(file_path: &str) -> Duration {
-    let audio_duration = anlg_audio_utils::source_from_path(file_path)
-        .ok()
-        .and_then(|source| source.total_duration());
-    direct_batch_timeout_for_audio(audio_duration)
-}
-
-fn direct_batch_timeout_for_audio(audio_duration: Option<Duration>) -> Duration {
-    let timeout = audio_duration
-        .map(|duration| {
-            duration
-                .saturating_mul(DIRECT_BATCH_AUDIO_DURATION_MULTIPLIER)
-                .saturating_add(DIRECT_BATCH_TIMEOUT_BUFFER)
-        })
-        .unwrap_or(DIRECT_BATCH_TIMEOUT_FLOOR);
-
-    timeout
-        .max(DIRECT_BATCH_TIMEOUT_FLOOR)
-        .min(DIRECT_BATCH_TIMEOUT_CEILING)
-}
-
-pub(super) async fn run_apple_speech_batch(
-    runtime: Arc<dyn BatchRuntime>,
-    params: BatchParams,
-    listen_params: owhisper_interface::ListenParams,
-) -> crate::Result<BatchRunOutput> {
-    let span = session_span(&params.session_id);
-
-    async {
-        let locale = anlg_transcribe_speechanalyzer::resolve_session_locale(
-            &listen_params.languages,
-        )
-        .ok_or_else(|| crate::BatchFailure::DirectRequestFailed {
-            provider: "apple-speech".to_string(),
-            message:
-                "Add this language in System Settings > General > Language & Region to transcribe it with Apple Speech."
-                    .to_string(),
-        })?;
-        let file_path = params.file_path.clone();
-        let started_at = Instant::now();
-
-        tracing::info!(
-            anarlog.stt.provider.name = "apple-speech",
-            anarlog.stt.language = %locale,
-            "apple_speech_batch_start"
-        );
-
-        let session_id = params.session_id.clone();
-        let progress_runtime = runtime.clone();
-        let transcribed = tokio::task::spawn_blocking(move || {
-            let progress = SoniqoProgressReporter {
-                runtime: progress_runtime,
-                session_id,
-            };
-            transcribe_apple_speech_file(&file_path, &locale, Some(&progress))
-        })
-        .await
-        .map_err(|e| crate::BatchFailure::DirectRequestFailed {
-            provider: "apple-speech".to_string(),
-            message: format!("Apple Speech transcription task failed: {e}"),
-        })?
-        .map_err(|e| {
-            let message = format_user_friendly_error(&e);
-            tracing::error!(
-                anarlog.stt.provider.name = "apple-speech",
-                error = %e,
-                anarlog.error.user_message = %message,
-                "apple_speech_batch_failed"
-            );
-            crate::BatchFailure::DirectRequestFailed {
-                provider: "apple-speech".to_string(),
-                message,
-            }
-        })?;
-
-        tracing::info!(
-            anarlog.stt.provider.name = "apple-speech",
-            elapsed_ms = started_at.elapsed().as_millis() as u64,
-            transcript.channel_count = transcribed.len(),
-            "apple_speech_batch_completed"
-        );
-
-        Ok(BatchRunOutput {
-            session_id: params.session_id,
-            mode: BatchRunMode::Direct,
-            response: anlg_transcribe_speechanalyzer::batch_response_from_transcripts(transcribed),
-        })
-    }
-    .instrument(span)
-    .await
-}
-
-/// Transcribes each recorded channel separately so mic and system audio stay attributable.
-fn transcribe_apple_speech_file(
-    file_path: &str,
-    locale: &str,
-    progress: Option<&SoniqoProgressReporter>,
-) -> std::result::Result<Vec<anlg_transcribe_speechanalyzer::FileTranscript>, String> {
-    let source = anlg_audio_utils::source_from_path(file_path).map_err(|e| e.to_string())?;
-    let channel_count = u16::from(source.channels()).max(1) as usize;
-    let samples =
-        anlg_audio_utils::resample_audio(source, TARGET_SAMPLE_RATE).map_err(|e| e.to_string())?;
-    let channels = collapse_identical_channels(split_resampled_channels(&samples, channel_count));
-
-    if let Some(progress) = progress {
-        progress.emit(SONIQO_PROGRESS_PLANNED);
-    }
-
-    let total = channels.len().max(1);
-    let mut transcripts = Vec::with_capacity(channels.len());
-
-    for (index, channel) in channels.into_iter().enumerate() {
-        let transcript = anlg_transcribe_speechanalyzer::transcribe_samples(&channel, locale)
-            .map_err(|e| e.to_string())?;
-        transcripts.push(transcript);
-
-        if let Some(progress) = progress {
-            progress.emit(soniqo_batch_progress(index + 1, total));
-        }
-    }
-
-    Ok(transcripts)
-}
-
-pub(super) async fn run_soniqo_batch(
-    runtime: Arc<dyn BatchRuntime>,
-    params: BatchParams,
-    listen_params: owhisper_interface::ListenParams,
-) -> crate::Result<BatchRunOutput> {
-    let span = session_span(&params.session_id);
-
-    async {
-        let model = listen_params
-            .model
-            .as_deref()
-            .ok_or_else(|| crate::BatchFailure::DirectRequestFailed {
-                provider: "soniqo".to_string(),
-                message: "Missing Soniqo model.".to_string(),
-            })?
-            .parse::<anlg_transcribe_soniqo::SoniqoModel>()
-            .map_err(|e| crate::BatchFailure::DirectRequestFailed {
-                provider: "soniqo".to_string(),
-                message: e.to_string(),
-            })?
-            .batch_model();
-
-        let file_path = params.file_path.clone();
-        let file_extension = Path::new(&file_path)
-            .extension()
-            .and_then(|extension| extension.to_str())
-            .unwrap_or_default()
-            .to_string();
-        let language = listen_params
-            .languages
-            .first()
-            .map(anlg_language::Language::bcp47_code);
-        let language_hint = soniqo_language_hint(language.as_deref());
-        let num_speakers = listen_params.num_speakers;
-        let language_label = language.as_deref().unwrap_or("auto").to_string();
-        let language_hint_label = language_hint.as_deref().unwrap_or("auto").to_string();
-        let started_at = Instant::now();
-
-        tracing::info!(
-            anarlog.stt.provider.name = "soniqo",
-            anarlog.stt.model = %model,
-            anarlog.stt.language = %language_label,
-            anarlog.stt.language_hint = %language_hint_label,
-            file.extension = %file_extension,
-            "soniqo_batch_start"
-        );
-
-        let session_id = params.session_id.clone();
-        let transcribed = tokio::task::spawn_blocking(move || {
-            let progress = SoniqoProgressReporter {
-                runtime,
-                session_id,
-            };
-            transcribe_soniqo_file(
-                model,
-                &file_path,
-                language_hint.as_deref(),
-                num_speakers,
-                Some(&progress),
-            )
-        })
-        .await
-        .map_err(|e| {
-            tracing::error!(
-                anarlog.stt.provider.name = "soniqo",
-                anarlog.stt.model = %model,
-                error = %e,
-                "soniqo_batch_task_join_failed"
-            );
-            crate::BatchFailure::DirectRequestFailed {
-                provider: "soniqo".to_string(),
-                message: format!("Soniqo transcription task failed: {e}"),
-            }
-        })?
-        .map_err(|e| {
-            let message = format_user_friendly_error(&e);
-            tracing::error!(
-                anarlog.stt.provider.name = "soniqo",
-                anarlog.stt.model = %model,
-                error = %e,
-                anarlog.error.user_message = %message,
-                "soniqo_batch_failed"
-            );
-            crate::BatchFailure::DirectRequestFailed {
-                provider: "soniqo".to_string(),
-                message,
-            }
-        })?;
-
-        tracing::info!(
-            anarlog.stt.provider.name = "soniqo",
-            anarlog.stt.model = %model,
-            elapsed_ms = started_at.elapsed().as_millis() as u64,
-            transcript.channel_count = transcribed.len(),
-            "soniqo_batch_completed"
-        );
-
-        let response = anlg_transcribe_soniqo::batch_response_from_channels(model, transcribed);
-
-        Ok(BatchRunOutput {
-            session_id: params.session_id,
-            mode: BatchRunMode::Direct,
-            response,
-        })
-    }
-    .instrument(span)
-    .await
-}
-
-fn transcribe_soniqo_file(
-    model: anlg_transcribe_soniqo::SoniqoModel,
-    file_path: &str,
-    language: Option<&str>,
-    num_speakers: Option<u32>,
-    progress: Option<&SoniqoProgressReporter>,
-) -> std::result::Result<Vec<anlg_transcribe_soniqo::FileTranscript>, String> {
-    let source = anlg_audio_utils::source_from_path(file_path).map_err(|e| e.to_string())?;
-    let channel_count = u16::from(source.channels()).max(1) as usize;
-    let sample_rate = u32::from(source.sample_rate());
-    let duration_ms = source
-        .total_duration()
-        .map(|duration| duration.as_millis().min(u128::from(u64::MAX)) as u64);
-
-    tracing::info!(
-        anarlog.stt.provider.name = "soniqo",
-        anarlog.stt.model = %model,
-        anarlog.stt.language = %language.unwrap_or("auto"),
-        audio.channel_count = channel_count,
-        audio.sample_rate_hz = sample_rate,
-        audio.duration_ms = duration_ms.unwrap_or_default(),
-        audio.duration_known = duration_ms.is_some(),
-        "soniqo_audio_file_loaded"
-    );
-
-    if channel_count <= 1 && !uses_resilient_soniqo_chunking(model) {
-        if let Some(progress) = progress {
-            progress.emit(SONIQO_PROGRESS_PLANNED);
-        }
-        tracing::info!(
-            anarlog.stt.provider.name = "soniqo",
-            anarlog.stt.model = %model,
-            "soniqo_single_channel_native_inference_start"
-        );
-        return anlg_transcribe_soniqo::transcribe_file(model, file_path, language)
-            .map(|transcript| {
-                if let Some(progress) = progress {
-                    progress.emit(SONIQO_PROGRESS_MAX);
-                }
-                vec![transcript]
-            })
-            .map_err(|e| e.to_string());
-    }
-
-    let resample_started_at = Instant::now();
-    let samples =
-        anlg_audio_utils::resample_audio(source, TARGET_SAMPLE_RATE).map_err(|e| e.to_string())?;
-    tracing::info!(
-        anarlog.stt.provider.name = "soniqo",
-        anarlog.stt.model = %model,
-        elapsed_ms = resample_started_at.elapsed().as_millis() as u64,
-        audio.source_sample_rate_hz = sample_rate,
-        audio.target_sample_rate_hz = TARGET_SAMPLE_RATE,
-        audio.resampled_sample_count = samples.len(),
-        "soniqo_audio_resampled"
-    );
-
-    let channel_samples =
-        collapse_identical_channels(split_resampled_channels(&samples, channel_count));
-    tracing::info!(
-        anarlog.stt.provider.name = "soniqo",
-        anarlog.stt.model = %model,
-        audio.source_channel_count = channel_count,
-        audio.transcribed_channel_count = channel_samples.len(),
-        "soniqo_channels_prepared"
-    );
-
-    let transcribed_channel_count = channel_samples.len();
-    let plans = channel_samples
-        .iter()
-        .enumerate()
-        .map(|(channel_index, samples)| {
-            soniqo_channel_plan(
-                model,
-                channel_index,
-                samples,
-                transcribed_channel_count == 2 && channel_index == 0,
-            )
-        })
-        .collect::<std::result::Result<Vec<_>, _>>()?;
-    let total_chunks = plans.iter().map(|plan| plan.chunks.len()).sum::<usize>();
-    let mut completed_chunks = 0usize;
-
-    if let Some(progress) = progress {
-        progress.emit(soniqo_batch_progress(0, total_chunks));
-    }
-
-    let mut transcripts = collect_soniqo_channel_transcripts(plans.into_iter().map(|plan| {
-        transcribe_soniqo_channel_chunks(model, plan, language, || {
-            completed_chunks += 1;
-            if let Some(progress) = progress {
-                progress.emit(soniqo_batch_progress(completed_chunks, total_chunks));
-            }
-        })
-    }))?;
-
-    for (channel_index, (transcript, samples)) in transcripts
-        .iter_mut()
-        .zip(channel_samples.iter())
-        .enumerate()
-    {
-        let Some(speaker_count) =
-            soniqo_diarization_speaker_count(num_speakers, channel_samples.len(), channel_index)
-        else {
-            continue;
-        };
-
-        let started_at = Instant::now();
-        match anlg_transcribe_soniqo::diarize_samples(model, samples, speaker_count) {
-            Ok(segments) => {
-                tracing::info!(
-                    anarlog.stt.provider.name = "soniqo",
-                    anarlog.stt.model = %model,
-                    channel.index = channel_index,
-                    speaker.count = speaker_count,
-                    segment.count = segments.len(),
-                    elapsed_ms = started_at.elapsed().as_millis() as u64,
-                    "soniqo_channel_diarization_completed"
-                );
-                transcript.speaker_segments = segments;
-            }
-            Err(error) => {
-                tracing::warn!(
-                    anarlog.stt.provider.name = "soniqo",
-                    anarlog.stt.model = %model,
-                    channel.index = channel_index,
-                    speaker.count = speaker_count,
-                    elapsed_ms = started_at.elapsed().as_millis() as u64,
-                    error = %error,
-                    "soniqo_channel_diarization_failed"
-                );
-            }
-        }
-    }
-
-    Ok(transcripts)
-}
-
-struct SoniqoProgressReporter {
-    runtime: Arc<dyn BatchRuntime>,
-    session_id: String,
-}
-
-impl SoniqoProgressReporter {
-    fn emit(&self, percentage: f64) {
-        self.runtime.emit(BatchEvent::BatchResponseStreamed {
-            session_id: self.session_id.clone(),
-            event: BatchStreamEvent::Progress {
-                percentage,
-                partial_text: None,
-            },
-        });
-    }
-}
-
-struct SoniqoChannelPlan {
-    channel_index: usize,
-    duration_seconds: f64,
-    is_direct_mic: bool,
-    chunks: Vec<AudioChunk>,
-}
-
-fn soniqo_language_hint(language: Option<&str>) -> Option<String> {
-    let language = language?.trim();
-    if language.is_empty() {
-        return None;
-    }
-
-    language
-        .split(['-', '_'])
-        .next()
-        .filter(|value| !value.is_empty())
-        .map(|value| value.to_lowercase())
-}
-
-fn uses_resilient_soniqo_chunking(model: anlg_transcribe_soniqo::SoniqoModel) -> bool {
-    matches!(model, anlg_transcribe_soniqo::SoniqoModel::ParakeetBatch)
-}
-
-fn soniqo_batch_progress(completed_chunks: usize, total_chunks: usize) -> f64 {
-    if total_chunks == 0 {
-        return SONIQO_PROGRESS_PLANNED;
-    }
-
-    let ratio = completed_chunks as f64 / total_chunks as f64;
-    (SONIQO_PROGRESS_PLANNED + ratio * SONIQO_PROGRESS_RANGE).min(SONIQO_PROGRESS_MAX)
-}
-
-fn soniqo_diarization_speaker_count(
-    num_speakers: Option<u32>,
-    channel_count: usize,
-    channel_index: usize,
-) -> Option<usize> {
-    let total = usize::try_from(num_speakers?).ok()?;
-
-    let count = match channel_count {
-        1 => total,
-        2 if channel_index == 1 => total.saturating_sub(1),
-        _ => return None,
-    };
-
-    (count >= 2).then_some(count)
-}
-
-fn collect_soniqo_channel_transcripts<I>(
-    transcripts: I,
-) -> std::result::Result<Vec<anlg_transcribe_soniqo::FileTranscript>, String>
-where
-    I: IntoIterator<Item = std::result::Result<anlg_transcribe_soniqo::FileTranscript, String>>,
-{
-    let mut output = Vec::new();
-    let mut successful_channels = 0usize;
-    let mut failed_channels = 0usize;
-
-    for transcript in transcripts {
-        match transcript {
-            Ok(transcript) => {
-                successful_channels += 1;
-                output.push(transcript);
-            }
-            Err(error) => {
-                failed_channels += 1;
-                tracing::warn!(
-                    anarlog.stt.provider.name = "soniqo",
-                    error = %error,
-                    "soniqo_channel_transcription_failed"
-                );
-                output.push(anlg_transcribe_soniqo::FileTranscript::new(
-                    String::new(),
-                    0.05,
-                ));
-            }
-        }
-    }
-
-    if successful_channels == 0 && failed_channels > 0 {
-        return Err(format!(
-            "Soniqo failed to transcribe all {failed_channels} audio channel(s)."
-        ));
-    }
-
-    Ok(output)
-}
-
-fn soniqo_channel_plan(
-    model: anlg_transcribe_soniqo::SoniqoModel,
-    channel_index: usize,
-    samples: &[f32],
-    is_direct_mic: bool,
-) -> std::result::Result<SoniqoChannelPlan, String> {
-    let duration_seconds = channel_duration_sec(samples);
-    let chunks = soniqo_channel_chunks(model, samples)?;
-    tracing::info!(
-        anarlog.stt.provider.name = "soniqo",
-        anarlog.stt.model = %model,
-        channel.index = channel_index,
-        channel.duration_seconds = duration_seconds,
-        channel.sample_count = samples.len(),
-        chunk.count = chunks.len(),
-        "soniqo_channel_chunked"
-    );
-
-    Ok(SoniqoChannelPlan {
-        channel_index,
-        duration_seconds,
-        is_direct_mic,
-        chunks,
-    })
-}
-
-fn transcribe_soniqo_channel_chunks(
-    model: anlg_transcribe_soniqo::SoniqoModel,
-    plan: SoniqoChannelPlan,
-    language: Option<&str>,
-    mut on_chunk_completed: impl FnMut(),
-) -> std::result::Result<anlg_transcribe_soniqo::FileTranscript, String> {
-    let mut texts = Vec::new();
-    let mut transcript_chunks = Vec::new();
-    let mut successful_chunks = 0usize;
-    let mut failed_chunks = 0usize;
-    let channel_index = plan.channel_index;
-    let is_direct_mic = plan.is_direct_mic;
-
-    for (chunk_index, chunk) in plan.chunks.into_iter().enumerate() {
-        let chunk_duration_ms =
-            (chunk.sample_end - chunk.sample_start) * 1000 / TARGET_SAMPLE_RATE as usize;
-        let chunk_rms = audio_rms(&chunk.samples);
-        if is_direct_mic && chunk_rms < SONIQO_DIRECT_MIC_MIN_RMS {
-            on_chunk_completed();
-            tracing::info!(
-                anarlog.stt.provider.name = "soniqo",
-                anarlog.stt.model = %model,
-                channel.index = channel_index,
-                chunk.index = chunk_index,
-                audio.rms = chunk_rms,
-                audio.minimum_rms = SONIQO_DIRECT_MIC_MIN_RMS,
-                "soniqo_direct_mic_chunk_skipped"
-            );
-            continue;
-        }
-
-        let chunk_started_at = Instant::now();
-        tracing::info!(
-            anarlog.stt.provider.name = "soniqo",
-            anarlog.stt.model = %model,
-            channel.index = channel_index,
-            chunk.index = chunk_index,
-            chunk.sample_start = chunk.sample_start,
-            chunk.sample_end = chunk.sample_end,
-            chunk.sample_count = chunk.samples.len(),
-            chunk.duration_ms = chunk_duration_ms,
-            "soniqo_chunk_native_inference_start"
-        );
-
-        let text = match transcribe_soniqo_samples(model, &chunk.samples, language) {
-            Ok(transcript) => {
-                successful_chunks += 1;
-                transcript.text
-            }
-            Err(e) => {
-                failed_chunks += 1;
-                tracing::warn!(
-                    anarlog.stt.provider.name = "soniqo",
-                    anarlog.stt.model = %model,
-                    channel.index = channel_index,
-                    chunk.index = chunk_index,
-                    elapsed_ms = chunk_started_at.elapsed().as_millis() as u64,
-                    error = %e,
-                    "soniqo_chunk_native_inference_failed"
-                );
-                on_chunk_completed();
-                continue;
-            }
-        };
-        on_chunk_completed();
-
-        tracing::info!(
-            anarlog.stt.provider.name = "soniqo",
-            anarlog.stt.model = %model,
-            channel.index = channel_index,
-            chunk.index = chunk_index,
-            elapsed_ms = chunk_started_at.elapsed().as_millis() as u64,
-            transcript.text_chars = text.chars().count(),
-            "soniqo_chunk_native_inference_completed"
-        );
-
-        let text = text.trim();
-        if !text.is_empty() {
-            texts.push(text.to_string());
-            transcript_chunks.push(anlg_transcribe_soniqo::FileTranscriptChunk {
-                text: text.to_string(),
-                start_seconds: chunk.sample_start as f64 / TARGET_SAMPLE_RATE as f64,
-                duration_seconds: (chunk.sample_end - chunk.sample_start) as f64
-                    / TARGET_SAMPLE_RATE as f64,
-            });
-        }
-    }
-
-    if successful_chunks == 0 && failed_chunks > 0 {
-        return Err(format!(
-            "Soniqo failed to transcribe all {failed_chunks} chunk(s) for channel {channel_index}."
-        ));
-    }
-
-    if failed_chunks > 0 {
-        tracing::warn!(
-            anarlog.stt.provider.name = "soniqo",
-            anarlog.stt.model = %model,
-            channel.index = channel_index,
-            chunk.success_count = successful_chunks,
-            chunk.failed_count = failed_chunks,
-            "soniqo_channel_completed_with_chunk_failures"
-        );
-    }
-
-    if transcript_chunks.is_empty() {
-        return Ok(anlg_transcribe_soniqo::FileTranscript::new(
-            texts.join(" "),
-            plan.duration_seconds,
-        ));
-    }
-
-    Ok(anlg_transcribe_soniqo::FileTranscript::from_chunks(
-        transcript_chunks,
-        plan.duration_seconds,
-    ))
-}
-
-fn audio_rms(samples: &[f32]) -> f64 {
-    if samples.is_empty() {
-        return 0.0;
-    }
-
-    let sum_of_squares = samples
-        .iter()
-        .map(|sample| f64::from(*sample).powi(2))
-        .sum::<f64>();
-    (sum_of_squares / samples.len() as f64).sqrt()
-}
-
-fn transcribe_soniqo_samples(
-    model: anlg_transcribe_soniqo::SoniqoModel,
-    samples: &[f32],
-    language: Option<&str>,
-) -> std::result::Result<anlg_transcribe_soniqo::FileTranscript, String> {
-    let file = tempfile::Builder::new()
-        .prefix("soniqo_channel_")
-        .suffix(".wav")
-        .tempfile()
-        .map_err(|e| e.to_string())?;
-    let spec = hound::WavSpec {
-        channels: 1,
-        sample_rate: TARGET_SAMPLE_RATE,
-        bits_per_sample: 32,
-        sample_format: hound::SampleFormat::Float,
-    };
-
-    {
-        let mut writer = hound::WavWriter::create(file.path(), spec).map_err(|e| e.to_string())?;
-        for sample in samples {
-            writer.write_sample(*sample).map_err(|e| e.to_string())?;
-        }
-        writer.finalize().map_err(|e| e.to_string())?;
-    }
-
-    anlg_transcribe_soniqo::transcribe_file(model, file.path(), language).map_err(|e| e.to_string())
-}
-
-fn soniqo_channel_chunks(
-    model: anlg_transcribe_soniqo::SoniqoModel,
-    samples: &[f32],
-) -> std::result::Result<Vec<AudioChunk>, String> {
-    if model == anlg_transcribe_soniqo::SoniqoModel::ParakeetBatch {
-        return Ok(split_audio_samples(
-            samples,
-            SONIQO_PARAKEET_MAX_CHUNK_SAMPLES,
-        ));
-    }
-
-    chunk_channel_audio::<anlg_audio_chunking::Error>(samples).map_err(|e| e.to_string())
-}
-
-fn split_audio_samples(samples: &[f32], max_samples: usize) -> Vec<AudioChunk> {
-    samples
-        .chunks(max_samples)
-        .enumerate()
-        .map(|(index, window)| {
-            let sample_start = index * max_samples;
-            let sample_end = sample_start + window.len();
-            AudioChunk {
-                samples: window.to_vec(),
-                sample_start,
-                sample_end,
-            }
-        })
-        .collect()
-}
-
-fn collapse_identical_channels(channels: Vec<Vec<f32>>) -> Vec<Vec<f32>> {
-    if channels.len() != 2 || !channels_are_effectively_identical(&channels[0], &channels[1]) {
-        return channels;
-    }
-
-    channels.into_iter().take(1).collect()
-}
-
-fn channels_are_effectively_identical(left: &[f32], right: &[f32]) -> bool {
-    if left.len().abs_diff(right.len()) > 1 {
-        return false;
-    }
-
-    let compared = left.len().min(right.len());
-    if compared == 0 {
-        return true;
-    }
-
-    let mean_abs_diff = left
-        .iter()
-        .zip(right.iter())
-        .map(|(a, b)| (a - b).abs())
-        .sum::<f32>()
-        / compared as f32;
-
-    mean_abs_diff < 0.0005
-}
+mod direct;
+mod local;
+
+pub(super) use direct::run_direct_batch_for_adapter_kind;
+pub(super) use local::{run_apple_speech_batch, run_soniqo_batch};
 
 #[cfg(test)]
-mod tests {
-    use std::future::{Future, pending};
-    use std::pin::Pin;
-    use std::sync::Arc;
-    use std::time::Duration;
+use direct::{
+    DIRECT_BATCH_TIMEOUT_CEILING, DIRECT_BATCH_TIMEOUT_FLOOR, direct_batch_timeout_for_audio,
+    prepare_anarlog_batch_upload, run_direct_batch_with_timeout,
+};
+#[cfg(test)]
+use local::{
+    SONIQO_DIRECT_MIC_MIN_RMS, SONIQO_PARAKEET_MAX_CHUNK_SAMPLES, SONIQO_PROGRESS_MAX,
+    SONIQO_PROGRESS_PLANNED, audio_rms, collapse_identical_channels,
+    collect_soniqo_channel_transcripts, soniqo_batch_progress, soniqo_channel_chunks,
+    soniqo_diarization_speaker_count, soniqo_language_hint, uses_resilient_soniqo_chunking,
+};
 
-    use axum::{Router, extract::State, http::StatusCode};
-    use owhisper_interface::batch::Response;
-    use tokio::sync::Notify;
-
-    use super::*;
-
-    #[derive(Clone, Default)]
-    struct HangingHttpAdapter;
-
-    impl BatchSttAdapter for HangingHttpAdapter {
-        fn provider_name(&self) -> &'static str {
-            "hanging-http"
-        }
-
-        fn is_supported_languages(
-            &self,
-            _languages: &[anlg_language::Language],
-            _model: Option<&str>,
-        ) -> bool {
-            true
-        }
-
-        fn transcribe_file<'a, P: AsRef<Path> + Send + 'a>(
-            &'a self,
-            client: &'a reqwest_middleware::ClientWithMiddleware,
-            api_base: &'a str,
-            _api_key: &'a str,
-            _params: &'a owhisper_interface::ListenParams,
-            _file_path: P,
-        ) -> Pin<
-            Box<
-                dyn Future<Output = std::result::Result<Response, owhisper_client::Error>>
-                    + Send
-                    + 'a,
-            >,
-        > {
-            Box::pin(async move {
-                client.post(api_base).body("audio").send().await?;
-                panic!("hanging provider unexpectedly responded");
-            })
-        }
-    }
-
-    #[derive(Clone)]
-    struct HangingProviderState {
-        request_started: Arc<Notify>,
-        request_cancelled: Arc<Notify>,
-    }
-
-    struct NotifyOnDrop(Arc<Notify>);
-
-    impl Drop for NotifyOnDrop {
-        fn drop(&mut self) {
-            self.0.notify_one();
-        }
-    }
-
-    async fn hanging_provider(State(state): State<HangingProviderState>) -> StatusCode {
-        let _cancelled = NotifyOnDrop(state.request_cancelled.clone());
-        state.request_started.notify_one();
-        pending::<()>().await;
-        StatusCode::OK
-    }
-
-    fn write_test_wav(path: &Path) {
-        let spec = hound::WavSpec {
-            channels: 1,
-            sample_rate: TARGET_SAMPLE_RATE,
-            bits_per_sample: 32,
-            sample_format: hound::SampleFormat::Float,
-        };
-        let mut writer = hound::WavWriter::create(path, spec).unwrap();
-        writer.write_sample(0.0f32).unwrap();
-        writer.finalize().unwrap();
-    }
-
-    fn write_test_wav_samples(path: &Path, samples: usize) {
-        let spec = hound::WavSpec {
-            channels: 1,
-            sample_rate: TARGET_SAMPLE_RATE,
-            bits_per_sample: 32,
-            sample_format: hound::SampleFormat::Float,
-        };
-        let mut writer = hound::WavWriter::create(path, spec).unwrap();
-        for _ in 0..samples {
-            writer.write_sample(0.0f32).unwrap();
-        }
-        writer.finalize().unwrap();
-    }
-
-    #[test]
-    fn direct_timeout_scales_with_audio_duration_and_is_bounded() {
-        assert_eq!(
-            direct_batch_timeout_for_audio(None),
-            DIRECT_BATCH_TIMEOUT_FLOOR
-        );
-        assert_eq!(
-            direct_batch_timeout_for_audio(Some(Duration::from_secs(60 * 60))),
-            Duration::from_secs(2 * 60 * 60 + 5 * 60)
-        );
-        assert_eq!(
-            direct_batch_timeout_for_audio(Some(Duration::from_secs(24 * 60 * 60))),
-            DIRECT_BATCH_TIMEOUT_CEILING
-        );
-    }
-
-    #[tokio::test]
-    async fn compresses_oversized_legacy_wav_before_anarlog_upload() {
-        let source = tempfile::Builder::new().suffix(".wav").tempfile().unwrap();
-        write_test_wav_samples(source.path(), TARGET_SAMPLE_RATE as usize);
-        let source_size = std::fs::metadata(source.path()).unwrap().len();
-        let max_bytes = source_size / 2;
-
-        let upload = prepare_anarlog_batch_upload(source.path().to_str().unwrap(), max_bytes)
-            .await
-            .unwrap();
-
-        assert_eq!(upload.path().extension().unwrap(), "mp3");
-        assert!(std::fs::metadata(upload.path()).unwrap().len() <= max_bytes);
-        assert!(source.path().exists());
-    }
-
-    #[tokio::test]
-    async fn keeps_anarlog_uploads_that_fit_without_reencoding() {
-        let source = tempfile::Builder::new().suffix(".wav").tempfile().unwrap();
-        write_test_wav(source.path());
-        let source_path = source.path().to_path_buf();
-        let max_bytes = std::fs::metadata(&source_path).unwrap().len();
-
-        let upload = prepare_anarlog_batch_upload(source_path.to_str().unwrap(), max_bytes)
-            .await
-            .unwrap();
-
-        assert_eq!(upload.path(), source_path);
-    }
-
-    #[tokio::test]
-    async fn direct_provider_timeout_cancels_non_responding_request() {
-        let request_started = Arc::new(Notify::new());
-        let request_cancelled = Arc::new(Notify::new());
-        let state = HangingProviderState {
-            request_started: request_started.clone(),
-            request_cancelled: request_cancelled.clone(),
-        };
-        let app = Router::new().fallback(hanging_provider).with_state(state);
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let address = listener.local_addr().unwrap();
-        let server = tokio::spawn(async move {
-            axum::serve(listener, app).await.unwrap();
-        });
-        let audio = tempfile::Builder::new().suffix(".wav").tempfile().unwrap();
-        write_test_wav(audio.path());
-        let params = BatchParams {
-            session_id: "timeout-test".to_string(),
-            provider: super::super::BatchProvider::OpenAI,
-            file_path: audio.path().to_string_lossy().into_owned(),
-            model: Some("whisper-1".to_string()),
-            base_url: format!("http://{address}/v1"),
-            api_key: "test".to_string(),
-            languages: vec![anlg_language::ISO639::En.into()],
-            keywords: vec![],
-            num_speakers: None,
-            min_speakers: None,
-            max_speakers: None,
-        };
-        let request = tokio::spawn(run_direct_batch_with_timeout::<HangingHttpAdapter>(
-            "hanging-http",
-            params,
-            owhisper_interface::ListenParams {
-                model: Some("whisper-1".to_string()),
-                channels: 1,
-                sample_rate: TARGET_SAMPLE_RATE,
-                languages: vec![anlg_language::ISO639::En.into()],
-                ..Default::default()
-            },
-            Duration::from_secs(5),
-        ));
-
-        tokio::time::timeout(Duration::from_secs(6), request_started.notified())
-            .await
-            .expect("provider did not receive the request");
-        let error = request
-            .await
-            .expect("batch task panicked")
-            .expect_err("non-responding provider should time out");
-
-        match error {
-            crate::Error::BatchFailed(failure) => {
-                assert_eq!(failure.code(), crate::BatchErrorCode::TimedOut);
-                assert!(matches!(
-                    failure,
-                    crate::BatchFailure::DirectRequestTimedOut { .. }
-                ));
-            }
-            other => panic!("unexpected timeout error: {other:?}"),
-        }
-        tokio::time::timeout(Duration::from_secs(2), request_cancelled.notified())
-            .await
-            .expect("provider request was not cancelled");
-
-        server.abort();
-    }
-
-    #[test]
-    fn collapses_effectively_identical_stereo_channels() {
-        let channels =
-            collapse_identical_channels(vec![vec![0.1, 0.2, 0.3], vec![0.1001, 0.2001, 0.3001]]);
-
-        assert_eq!(channels, vec![vec![0.1, 0.2, 0.3]]);
-    }
-
-    #[test]
-    fn keeps_distinct_stereo_channels() {
-        let channels = collapse_identical_channels(vec![vec![0.1, 0.2], vec![0.9, 0.8]]);
-
-        assert_eq!(channels, vec![vec![0.1, 0.2], vec![0.9, 0.8]]);
-    }
-
-    #[test]
-    fn parakeet_batch_uses_fixed_audio_windows() {
-        let samples =
-            vec![0.0; SONIQO_PARAKEET_MAX_CHUNK_SAMPLES * 2 + TARGET_SAMPLE_RATE as usize];
-        let chunks =
-            soniqo_channel_chunks(anlg_transcribe_soniqo::SoniqoModel::ParakeetBatch, &samples)
-                .unwrap();
-
-        assert_eq!(chunks.len(), 3);
-        assert_eq!(chunks[0].sample_start, 0);
-        assert_eq!(chunks[0].sample_end, SONIQO_PARAKEET_MAX_CHUNK_SAMPLES);
-        assert_eq!(chunks[1].sample_start, chunks[0].sample_end);
-        assert_eq!(chunks[1].sample_end, SONIQO_PARAKEET_MAX_CHUNK_SAMPLES * 2);
-        assert_eq!(chunks[2].sample_start, chunks[1].sample_end);
-        assert_eq!(chunks[2].samples.len(), TARGET_SAMPLE_RATE as usize);
-    }
-
-    #[test]
-    fn direct_mic_gate_rejects_only_quiet_chunks() {
-        assert!(audio_rms(&[0.0007; 100]) < SONIQO_DIRECT_MIC_MIN_RMS);
-        assert!(audio_rms(&[0.0009; 100]) >= SONIQO_DIRECT_MIC_MIN_RMS);
-    }
-
-    #[test]
-    fn parakeet_batch_window_bounds_force_coreml_shape_3000() {
-        let minimum_mel_frames = TARGET_SAMPLE_RATE as usize * 20 / 160 + 1;
-        let maximum_mel_frames = SONIQO_PARAKEET_MAX_CHUNK_SAMPLES / 160 + 1;
-
-        assert!(minimum_mel_frames > 2000);
-        assert!(maximum_mel_frames <= 3000);
-    }
-
-    #[test]
-    fn soniqo_language_hint_uses_base_language_code() {
-        assert_eq!(soniqo_language_hint(Some("de-DE")).as_deref(), Some("de"));
-        assert_eq!(soniqo_language_hint(Some("en_US")).as_deref(), Some("en"));
-        assert_eq!(soniqo_language_hint(Some(" fr ")).as_deref(), Some("fr"));
-        assert_eq!(soniqo_language_hint(Some("")).as_deref(), None);
-        assert_eq!(soniqo_language_hint(None).as_deref(), None);
-    }
-
-    #[test]
-    fn parakeet_batch_uses_resilient_chunking() {
-        assert!(uses_resilient_soniqo_chunking(
-            anlg_transcribe_soniqo::SoniqoModel::ParakeetBatch
-        ));
-        assert!(!uses_resilient_soniqo_chunking(
-            anlg_transcribe_soniqo::SoniqoModel::Omnilingual
-        ));
-    }
-
-    #[test]
-    fn soniqo_diarization_uses_remote_count_for_system_channel() {
-        assert_eq!(soniqo_diarization_speaker_count(Some(3), 2, 0), None);
-        assert_eq!(soniqo_diarization_speaker_count(Some(3), 2, 1), Some(2));
-    }
-
-    #[test]
-    fn soniqo_diarization_uses_total_count_for_mono_audio() {
-        assert_eq!(soniqo_diarization_speaker_count(Some(2), 1, 0), Some(2));
-        assert_eq!(soniqo_diarization_speaker_count(Some(1), 1, 0), None);
-        assert_eq!(soniqo_diarization_speaker_count(None, 1, 0), None);
-    }
-
-    #[test]
-    fn soniqo_progress_starts_after_chunk_planning() {
-        assert_eq!(soniqo_batch_progress(0, 10), SONIQO_PROGRESS_PLANNED);
-        assert_eq!(soniqo_batch_progress(0, 0), SONIQO_PROGRESS_PLANNED);
-    }
-
-    #[test]
-    fn soniqo_progress_caps_before_completion() {
-        assert!((soniqo_batch_progress(5, 10) - 0.5).abs() < 1e-9);
-        assert_eq!(soniqo_batch_progress(10, 10), SONIQO_PROGRESS_MAX);
-        assert_eq!(soniqo_batch_progress(11, 10), SONIQO_PROGRESS_MAX);
-    }
-
-    #[test]
-    fn collect_soniqo_channel_transcripts_keeps_channel_slots() {
-        let transcripts = collect_soniqo_channel_transcripts([
-            Ok(anlg_transcribe_soniqo::FileTranscript::new(
-                "hello".to_string(),
-                1.0,
-            )),
-            Err("native chunk failed".to_string()),
-        ])
-        .unwrap();
-
-        assert_eq!(transcripts.len(), 2);
-        assert_eq!(transcripts[0].text, "hello");
-        assert_eq!(transcripts[1].text, "");
-    }
-
-    #[test]
-    fn collect_soniqo_channel_transcripts_preserves_later_channel_index() {
-        let transcripts = collect_soniqo_channel_transcripts([
-            Err("first failed".to_string()),
-            Ok(anlg_transcribe_soniqo::FileTranscript::new(
-                "system audio".to_string(),
-                1.0,
-            )),
-        ])
-        .unwrap();
-
-        let response = anlg_transcribe_soniqo::batch_response_from_channels(
-            anlg_transcribe_soniqo::SoniqoModel::ParakeetBatch,
-            transcripts,
-        );
-        let alternative = &response.results.channels[1].alternatives[0];
-
-        assert_eq!(response.results.channels.len(), 2);
-        assert_eq!(response.results.channels[0].alternatives[0].transcript, "");
-        assert_eq!(alternative.transcript, "system audio");
-        assert_eq!(alternative.words[0].channel, 1);
-    }
-
-    #[test]
-    fn collect_soniqo_channel_transcripts_errors_when_all_channels_fail() {
-        let error = collect_soniqo_channel_transcripts([
-            Err("first failed".to_string()),
-            Err("second failed".to_string()),
-        ])
-        .unwrap_err();
-
-        assert_eq!(error, "Soniqo failed to transcribe all 2 audio channel(s).");
-    }
-}
+#[cfg(test)]
+mod tests;
